Refactor payload validation exception handling
Moved PayloadValidationException and related types from the API app to the application-generic library for reuse. Updated trigger-event usecase to handle payload validation and apply defaults, while removing redundant validation logic from parse-event-request usecase.
